refactor: extract app actions and import flow into sibling files
Move the four app-level GAction/dialog handlers (setup_gactions,
show_shortcuts, show_about, show_preferences) into actions.rs and the
folder-import entry points (show_import_dialog, run_import) plus their
GIO helper functions (resolve_folder_via_gio, gio_walk) into import.rs.
Pure code movement to shrink application/mod.rs — no behavior changes.
